feat: divide store file
