+++ src/store/modules/todos/actions.js
@@ -1,0 +1,16 @@
+export const actions = {
+  complete(context, payload) {
+    context.commit('completeTodo', payload);
+  },
+
+  init({
+    commit
+  }, payload) {
+    setTimeout(() => {
+      commit({
+        type: 'initTodo',
+        id: payload.id
+      });
+    }, 1000)
+  }
+};
