+++ src/store/modules/todos/getters.js
@@ -1,0 +1,8 @@
+export const getters = {
+  completedTodos(state) {
+    return state.todos.filter(todo => todo.completed);
+  },
+  completedTodosCount(state, getters) {
+    return getters.completedTodos.length;
+  }
+};
