+++ src/store/modules/todos/index.js
@@ -1,0 +1,19 @@
+import {
+  state
+} from './state';
+import {
+  getters
+} from './getters';
+import {
+  mutations
+} from './mutations';
+import {
+  actions
+} from './actions';
+
+export {
+  state,
+  getters,
+  mutations,
+  actions
+};
